Create local package and extract methods to Parser and Mapper class
